Move balance cell animation logic from adapter into holder
